# backend/backend/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Cash Platform API
  version: 0.1.0
servers:
  - url: http://localhost:3000
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
            details:
              type: array
              items:
                type: object
                properties:
                  path:
                    type: string
                  message:
                    type: string
    VersionInfo:
      type: object
      properties:
        name:
          type: string
        version:
          type: string
        gitSha:
          type: string
        gitBranch:
          type: string
        buildTime:
          type: string
    AuthLoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    AuthLoginResponse:
      type: object
      properties:
        token:
          type: string
    Account:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        balance:
          type: string
        userId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AccountCreateRequest:
      type: object
      required: [name, balance]
      properties:
        name:
          type: string
        balance:
          type: string
    Transaction:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
        amount:
          type: string
        fromAccountId:
          type: string
          format: uuid
        toAccountId:
          type: string
          format: uuid
        categoryId:
          type: string
          format: uuid
          nullable: true
        description:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TransactionCreateRequest:
      type: object
      required: [type, amount]
      properties:
        type:
          type: string
          example: transfer
        amount:
          type: string
        fromAccountId:
          type: string
          format: uuid
        toAccountId:
          type: string
          format: uuid
        categoryId:
          type: string
          format: uuid
        description:
          type: string
    Category:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        userId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CategoryCreateRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string

paths:
  /health:
    get:
      summary: Health check
      responses:
        "200":
          description: OK

  /version:
    get:
      summary: Version info
      responses:
        "200":
          description: Version metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VersionInfo"

  /metrics:
    get:
      summary: Prometheus metrics
      responses:
        "200":
          description: Prometheus metrics

  /auth/login:
    post:
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthLoginRequest"
      responses:
        "200":
          description: JWT token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthLoginResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Too many requests
        "500":
          description: Server error

  /accounts:
    get:
      summary: List accounts
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
        "401":
          description: Unauthorized
        "429":
          description: Too many requests
        "500":
          description: Server error
    post:
      summary: Create account
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
        "429":
          description: Too many requests
        "500":
          description: Server error

  /transactions:
    get:
      summary: List transactions
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
        "401":
          description: Unauthorized
        "429":
          description: Too many requests
        "500":
          description: Server error
    post:
      summary: Create transaction
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
        "429":
          description: Too many requests
        "500":
          description: Server error

  /categories:
    get:
      summary: List categories
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
        "401":
          description: Unauthorized
        "429":
          description: Too many requests
        "500":
          description: Server error
    post:
      summary: Create category
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
        "429":
          description: Too many requests
        "500":
          description: Server error